taskwave_backend_ready/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary]>=3.2
pydantic>=2.8
pydantic-settings>=2.4
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
alembic>=1.13
email-validator>=2.1
boto3>=1.34
